# swagger.yaml
swagger: "2.0"
info:
  description: "here is my Restuarant service api documentation."
  version: "1.0.0"
  title: "Swagger Restuarant"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8082"
basePath: "/"
schemes:
- "http"
- "https"
paths:   
...
  /restuarants/update-restuarant/{id}:  
    put:
      tags:
      - "Restuarant"
      summary: "Update an existing Restuarant"
      description: ""
      operationId: "updateRestuarant"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Pet"
      responses:
        "400":
          description: ""
        "404":
          description: ""
        "405":
          description: ""
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
...
  /restuarants/update-menu/{id}:  
    put:
      tags:
      - "Restuarant"
      summary: "Update an existing Restuarant"
      description: ""
      operationId: "updateRestuarant"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/Pet"
      responses:
        "400":
          description: ""
        "404":
          description: ""
        "405":
          description: ""
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
...